decompose passage mutation
